🏗️ Move components related to comment feature into new dedicated folder
