+++ src/features/activity-feed/comment/components/Comment/style.scss
@@ -1,5 +1,5 @@
-@use '../../styles/settings/breakpoints' as b;
-@use '../../styles/settings/colors' as c;
+@use '../../../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/colors' as c;
 
 .c-comment-list {
     padding-left: 0 !important;
