+++ src/features/activity-feed/comment/components/CommentForm/index.jsx
@@ -3,7 +3,7 @@ import { useForm } from 'react-hook-form'
 import { useDispatch } from 'react-redux'
 import { Paper, InputBase, IconButton } from '@mui/material'
 import SendIcon from '@mui/icons-material/Send'
-import { createComment } from '../../../redux/thunks/feed'
+import { createComment } from '../../../../../redux/thunks/feed'
 import './style.scss'
 
 CommentForm.propTypes = {
