+++ src/features/activity-feed/post/components/Post/index.jsx
@@ -10,8 +10,8 @@ import pluralize from 'pluralize'
 import { getPostComments, getPostReactions } from '../../../../../redux/selectors/feed'
 import { fetchComments } from '../../../../../redux/thunks/feed'
 import { getUser } from '../../../../user/store/userSelectors'
-import { CommentForm } from '../../../../../components/Forms/CommentForm'
-import { Comment } from '../../../../../components/Comment'
+import { CommentForm } from '../../../comment/components/CommentForm'
+import { Comment } from '../../../comment/components/Comment'
 import { ReactionButton } from '../../../../../components/Buttons/ReactionButton'
 import { PostReactionsCounter } from '../../../../../components/PostReactionsCounter'
